+++ src/Builder/UuidBuilderInterface.php
@@ -28,10 +28,10 @@ interface UuidBuilderInterface
      * Builds and returns a UuidInterface
      *
      * @param CodecInterface $codec The codec to use for building this UuidInterface instance
-     * @param string[] $fields An array of fields from which to construct a UuidInterface instance
+     * @param string $bytes The byte string from which to construct a UUID
      *
      * @return UuidInterface Implementations may choose to return more specific
      *     instances of UUIDs that implement UuidInterface
      */
-    public function build(CodecInterface $codec, array $fields): UuidInterface;
+    public function build(CodecInterface $codec, string $bytes): UuidInterface;
 }