+++ tests/Encoder/TimestampFirstCombCodecTest.php
@@ -63,14 +63,14 @@ class TimestampFirstCombCodecTest extends TestCase
             ->method('build')
             ->with(
                 $this->codec,
-                [
+                hex2bin(implode('', [
                     'time_low' => 'ff6f8cb0',
                     'time_mid' => 'c57d',
                     'time_hi_and_version' => '11e1',
                     'clock_seq_hi_and_reserved' => '9b',
                     'clock_seq_low' => '21',
                     'node' => '0800200c9a66',
-                ]
+                ]))
             );
         $this->codec->decode('0800200c-9a66-11e1-9b21-ff6f8cb0c57d');
     }
@@ -81,14 +81,14 @@ class TimestampFirstCombCodecTest extends TestCase
             ->method('build')
             ->with(
                 $this->codec,
-                [
+                hex2bin(implode('', [
                     'time_low' => 'ff6f8cb0',
                     'time_mid' => 'c57d',
                     'time_hi_and_version' => '11e1',
                     'clock_seq_hi_and_reserved' => '9b',
                     'clock_seq_low' => '21',
                     'node' => '0800200c9a66',
-                ]
+                ]))
             );
         $this->codec->decodeBytes((string) hex2bin('0800200c9a6611e19b21ff6f8cb0c57d'));
     }